extracted few things from lib.py
