+++ pype/lib/__init__.py
@@ -1,0 +1,11 @@
+# -*- coding: utf-8 -*-
+"""Pype lib module."""
+from .hooks import PypeHook, execute_hook
+from .plugin_tools import filter_pyblish_plugins
+
+__all__ = [
+    "PypeHook",
+    "execute_hook",
+
+    "filter_pyblish_plugins"
+]
